# tests/data/openapi/agent.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "MeResponse": {
        "properties": {
          "avatar_preset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Preset"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "onboarded_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarded At"
          },
          "questions_asked": {
            "default": 0,
            "title": "Questions Asked",
            "type": "integer"
          },
          "role": {
            "anyOf": [
              {
                "enum": [
                  "lawyer",
                  "accountant",
                  "manager",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "tasks": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "law_questions",
                    "contract_review",
                    "drafting",
                    "files"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tasks"
          },
          "tour_completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tour Completed At"
          },
          "usage_kind": {
            "anyOf": [
              {
                "enum": [
                  "personal",
                  "business"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage Kind"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "ProfileUpdateRequest": {
        "description": "Partial profile update: only the fields present in the request body\nare applied (`exclude_unset` semantics in the route). `onboarded: true`\nand `tour_completed: true` stamp their timestamps once — repeats are\nidempotent.",
        "properties": {
          "avatar_preset": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Preset"
          },
          "first_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "onboarded": {
            "default": false,
            "title": "Onboarded",
            "type": "boolean"
          },
          "role": {
            "anyOf": [
              {
                "enum": [
                  "lawyer",
                  "accountant",
                  "manager",
                  "other"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "tasks": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "law_questions",
                    "contract_review",
                    "drafting",
                    "files"
                  ],
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tasks"
          },
          "tour_completed": {
            "default": false,
            "title": "Tour Completed",
            "type": "boolean"
          },
          "usage_kind": {
            "anyOf": [
              {
                "enum": [
                  "personal",
                  "business"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usage Kind"
          }
        },
        "title": "ProfileUpdateRequest",
        "type": "object"
      },
...
      "RegisterRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
      "SourceArticleDetail": {
        "properties": {
          "act_short_name": {
            "title": "Act Short Name",
            "type": "string"
          },
          "article_id": {
            "format": "uuid",
            "title": "Article Id",
            "type": "string"
          },
          "full_text": {
            "title": "Full Text",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "article_id",
          "act_short_name",
          "number",
          "full_text"
        ],
        "title": "SourceArticleDetail",
        "type": "object"
      },
      "SourceArticleListItem": {
        "properties": {
          "article_id": {
            "format": "uuid",
            "title": "Article Id",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "article_id",
          "number"
        ],
        "title": "SourceArticleListItem",
        "type": "object"
      },
      "SourceArticlesResponse": {
        "properties": {
          "articles": {
            "items": {
              "$ref": "#/components/schemas/SourceArticleListItem"
            },
            "title": "Articles",
            "type": "array"
          },
          "short_name": {
            "title": "Short Name",
            "type": "string"
          },
          "source_doc_id": {
            "title": "Source Doc Id",
            "type": "string"
          }
        },
        "required": [
          "source_doc_id",
          "short_name",
          "articles"
        ],
        "title": "SourceArticlesResponse",
        "type": "object"
      },
      "SourceSummary": {
        "properties": {
          "branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "redaction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redaction"
          },
          "short_name": {
            "title": "Short Name",
            "type": "string"
          },
          "source_doc_id": {
            "title": "Source Doc Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "in_corpus",
              "planned"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "source_doc_id",
          "short_name",
          "full_name",
          "kind",
          "branch",
          "redaction",
          "status"
        ],
        "title": "SourceSummary",
        "type": "object"
      },
      "SourcesResponse": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceSummary"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "sources"
        ],
        "title": "SourcesResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Neurolegal Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/me": {
      "get": {
        "operationId": "me_auth_me_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "neurolegal_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Neurolegal Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/profile": {
      "patch": {
        "operationId": "patch_profile_profile_patch",
        "parameters": [
          {
            "in": "cookie",
            "name": "neurolegal_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Neurolegal Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Profile",
        "tags": [
          "profile"
        ]
      }
    },
...
    "/sources": {
      "get": {
        "operationId": "sources_sources_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "neurolegal_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Neurolegal Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sources"
      }
    },
    "/sources/articles/{article_id}": {
      "get": {
        "operationId": "article_sources_articles__article_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "article_id",
            "required": true,
            "schema": {
              "title": "Article Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "neurolegal_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Neurolegal Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceArticleDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Article"
      }
    },
    "/sources/{source_doc_id}/articles": {
      "get": {
        "operationId": "source_articles_sources__source_doc_id__articles_get",
        "parameters": [
          {
            "in": "path",
            "name": "source_doc_id",
            "required": true,
            "schema": {
              "title": "Source Doc Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "neurolegal_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Neurolegal Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceArticlesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Source Articles"
      }
    },
...
  }
}